Caas vlan test added
