+++ PowershellModule/CloudCompute.Powershell.Tests/MockApis/GET/caas/2.3/_/network/vlan/_q/id=91d21154-3414-4f87-9af5-a656884b49bc
@@ -1,0 +1,1 @@
+{"Method":"GET","Uri":"https://api-au.dimensiondata.com/caas/2.3/1578108f-e4aa-4ab7-8b2b-b9244482df93/network/vlan?id=91d21154-3414-4f87-9af5-a656884b49bc&","StatusCode":"OK","RequestContent":"","ResponseContent":"<?xml version=\"1.0\" encoding=\"UTF-8\"?><vlans xmlns=\"urn:didata.com:api:cloud:types\" pageNumber=\"1\" pageCount=\"1\" totalCount=\"1\" pageSize=\"250\"><vlan id=\"91d21154-3414-4f87-9af5-a656884b49bc\" datacenterId=\"AU9\"><networkDomain id=\"7782f3c3-5482-4db6-90e9-15e5e80fb440\" name=\"EMaaS_Client_Test\"/><name>Prod</name><description></description><privateIpv4Range address=\"10.0.0.0\" prefixSize=\"24\"/><gatewayAddressing>LOW</gatewayAddressing><ipv4GatewayAddress>10.0.0.1</ipv4GatewayAddress><ipv6Range address=\"2402:9900:111:1110:0:0:0:0\" prefixSize=\"64\"/><ipv6GatewayAddress>2402:9900:111:1110:0:0:0:1</ipv6GatewayAddress><createTime>2015-10-22T23:57:41.000Z</createTime><state>NORMAL</state></vlan></vlans>","ResponseContentType":"application/xml","Mock":true}
