+++ PowershellModule/CloudCompute.Powershell.Tests/MockApis/GET/caas/2.3/_/network/vlan/_q/id=c0512714-6be2-467e-a123-98959679365c
@@ -1,0 +1,1 @@
+{"Method":"GET","Uri":"https://api-au.dimensiondata.com/caas/2.3/1578108f-e4aa-4ab7-8b2b-b9244482df93/network/vlan?id=c0512714-6be2-467e-a123-98959679365c&","StatusCode":"OK","RequestContent":"","ResponseContent":"<?xml version=\"1.0\" encoding=\"UTF-8\"?><vlans xmlns=\"urn:didata.com:api:cloud:types\" pageNumber=\"1\" pageCount=\"0\" totalCount=\"0\" pageSize=\"250\"/>","ResponseContentType":"application/xml","Mock":true}
