+++ PowershellModule/CloudCompute.Powershell.Tests/MockApis/GET/caas/2.3/_/network/vlan/_q/name=vlan01
@@ -1,0 +1,1 @@
+{"Method":"GET","Uri":"https://api-au.dimensiondata.com/caas/2.3/1578108f-e4aa-4ab7-8b2b-b9244482df93/network/vlan?name=vlan01&","StatusCode":"OK","RequestContent":"","ResponseContent":"<?xml version=\"1.0\" encoding=\"UTF-8\"?><vlans xmlns=\"urn:didata.com:api:cloud:types\" pageNumber=\"1\" pageCount=\"1\" totalCount=\"1\" pageSize=\"250\"><vlan id=\"b901038f-0b0b-49fa-b55c-ad6752cb7096\" datacenterId=\"AU9\"><networkDomain id=\"0fdb588c-867f-42ef-b839-16f518ae4f6a\" name=\"River_Lab\"/><name>vlan01</name><privateIpv4Range address=\"172.20.0.0\" prefixSize=\"24\"/><gatewayAddressing>LOW</gatewayAddressing><ipv4GatewayAddress>172.20.0.1</ipv4GatewayAddress><ipv6Range address=\"2402:9900:111:1439:0:0:0:0\" prefixSize=\"64\"/><ipv6GatewayAddress>2402:9900:111:1439:0:0:0:1</ipv6GatewayAddress><createTime>2016-08-16T23:53:28.000Z</createTime><state>NORMAL</state></vlan></vlans>","ResponseContentType":"application/xml","Mock":true}
