+++ PowershellModule/CloudCompute.Powershell.Tests/MockApis/GET/caas/2.3/_/network/vlan/_q/networkDomainId=bcf42c64-8d19-4844-b0ae-eac6a2f4b486
@@ -1,0 +1,1 @@
+{"Method":"GET","Uri":"https://api-au.dimensiondata.com/caas/2.3/1578108f-e4aa-4ab7-8b2b-b9244482df93/network/vlan?networkDomainId=bcf42c64-8d19-4844-b0ae-eac6a2f4b486&","StatusCode":"OK","RequestContent":"","ResponseContent":"<?xml version=\"1.0\" encoding=\"UTF-8\"?><vlans xmlns=\"urn:didata.com:api:cloud:types\" pageNumber=\"1\" pageCount=\"0\" totalCount=\"0\" pageSize=\"250\"/>","ResponseContentType":"application/xml","Mock":true}
