+++ PowershellModule/CloudCompute.Powershell.Tests/ServerTests/NetworkDomain.Tests.ps1
@@ -67,9 +67,9 @@ Describe "Remove-CaaSNetworkDomain" {
 		$networkIdInfo.value = "a4f484de-b9ed-43e4-b565-afbf69417615"
 		$response.info += $networkIdInfo
 		$response.responseCode = "OK"
-		$testConnection | Setup "POST" "/caas/2.1/$($testConnection.CaaSClientId)/network/deployNetworkDomain" $response 200
-		New-CaasNetworkDomain -Connection $testConnection.CaaSConnection -DatacenterId AU9 -Name RandomNetworkName_1 -Type Advanced
-		$testConnection | Verify "POST" "/caas/2.1/$($testConnection.CaaSClientId)/network/deployNetworkDomain" 1
+		$testConnection | Setup "POST" "/caas/2.1/$($testConnection.CaaSClientId)/network/deleteNetworkDomain" $response 200
+		Remove-CaasNetworkDomain -Connection $testConnection.CaaSConnection -NetworkDomain $domains
+		$testConnection | Verify "POST" "/caas/2.1/$($testConnection.CaaSClientId)/network/deleteNetworkDomain" 1
 	}
 }
 
@@ -92,24 +92,62 @@ Describe "Get-CaasVlan By Datacenter Id" {
 Describe "Get-CaasVlan filter by Vlan Name" {
     It "Get-CaasVlan Should Get the list of Vlan by name" {
 		$testConnection = New-CaaSTestConnection -TestContext $TestContext
-		$vlans = Get-CaasVlan -Connection $testConnection.CaaSConnection -Name River_vlan
-		$testConnection | Verify "GET" "/caas/2.3/$($testConnection.CaaSClientId)/network/vlan?name=River_vlan" 1
+		$vlans = Get-CaasVlan -Connection $testConnection.CaaSConnection -Name vlan01
+		$testConnection | Verify "GET" "/caas/2.3/$($testConnection.CaaSClientId)/network/vlan?name=vlan01" 1
 	}
 }
 
 Describe "Get-CaasVlan By Vlan Id" {
     It "List Vlan Should have id filter" {
 		$testConnection = New-CaaSTestConnection -TestContext $TestContext
-		$vlans = Get-CaasVlan -Connection $testConnection.CaaSConnection -Id c0512714-6be2-467e-a123-98959679365c
-		$testConnection | Verify "GET" "/caas/2.3/$($testConnection.CaaSClientId)/network/vlan?id=c0512714-6be2-467e-a123-98959679365c" 1
+		$vlans = Get-CaasVlan -Connection $testConnection.CaaSConnection -Id 91d21154-3414-4f87-9af5-a656884b49bc
+		$testConnection | Verify "GET" "/caas/2.3/$($testConnection.CaaSClientId)/network/vlan?id=91d21154-3414-4f87-9af5-a656884b49bc" 1
 	}
 }
 
 Describe "Get-CaasVlan By Network Domain" {
     It "List Vlan Should have id filter" {
 		$testConnection = New-CaaSTestConnection -TestContext $TestContext
-		$networkdomain = Get-CaaSNetworkDomain -Connection $testConnection.CaaSConnection -Name River_Lab
+		$networkdomain = Get-CaaSNetworkDomain -Connection $testConnection.CaaSConnection -Id 0fdb588c-867f-42ef-b839-16f518ae4f6a
 		$vlans = Get-CaasVlan -Connection $testConnection.CaaSConnection -NetworkDomain $networkdomain
 		$testConnection | Verify "GET" "/caas/2.3/$($testConnection.CaaSClientId)/network/vlan?networkDomainid=$($networkdomain.id)" 1
+	}
+}
+
+Describe "New-CaasVlan" {
+    It "New-CaasVlan Should Create a new Vlan" {
+		$testConnection = New-CaaSTestConnection -TestContext $TestContext
+		$networkdomain = Get-CaaSNetworkDomain -Connection $testConnection.CaaSConnection -Name River_Lab
+		$response = new-object -TypeName  DD.CBU.Compute.Api.Contracts.Network20.ResponseType
+		$response.operation = "CREATE VLAN"
+		$response.message = "CREATE VLAN"
+		$response.info = @()
+		$vlanInfo = new-object -TypeName DD.CBU.Compute.Api.Contracts.Network20.NameValuePairType
+		$vlanInfo.name = "vlanId"
+		$vlanInfo.value = "1578108f-e4aa-4ab7-8b2b-b9244482df93"
+		$response.info += $vlanInfo
+		$response.responseCode = "OK"
+		$testConnection | Setup "POST" "/caas/2.3/$($testConnection.CaaSClientId)/network/deployVlan" $response 200
+		New-CaasVlan -Connection $testConnection.CaaSConnection -NetworkDomain $networkdomain -Name vlan02 -PrivateIpv4BaseAddress 172.20.2.0 -PrivateIpv4PrefixSize 24
+		$testConnection | Verify "POST" "/caas/2.3/$($testConnection.CaaSClientId)/network/deployVlan" 1
+	}
+}
+
+Describe "Remove-CaasVlan" {
+    It "Remove-CaasVlan Should remove a Vlan" {
+		$testConnection = New-CaaSTestConnection -TestContext $TestContext
+		$vlans = Get-CaasVlan -Connection $testConnection.CaaSConnection -Id 91d21154-3414-4f87-9af5-a656884b49bc
+		$response = new-object -TypeName  DD.CBU.Compute.Api.Contracts.Network20.ResponseType
+		$response.operation = "REMOVE VLAN"
+		$response.message = "REMOVE VLAN"
+		$response.info = @()
+		$vlanInfo = new-object -TypeName DD.CBU.Compute.Api.Contracts.Network20.NameValuePairType
+		$vlanInfo.name = "vlanId"
+		$vlanInfo.value = "1578108f-e4aa-4ab7-8b2b-b9244482df93"
+		$response.info += $vlanInfo
+		$response.responseCode = "OK"
+		$testConnection | Setup "POST" "/caas/2.3/$($testConnection.CaaSClientId)/network/deleteVlan" $response 200
+		Remove-CaasVlan -Connection $testConnection.CaaSConnection -Vlan $vlans
+		$testConnection | Verify "POST" "/caas/2.3/$($testConnection.CaaSClientId)/network/deleteVlan" 1
 	}
 }
